fix: move util function to single-line-regex-linter.spec.ts
